# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Comerc API Documentation",
        "description": "Documentação da API Comerc",
        "contact": {
            "email": "contato@leonardo-marques.tech"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Lista todos os clientes",
                "description": "Display a listing of the resource.",
                "operationId": "413c12320d4121431e38339f45ecaa7c",
                "responses": {
                    "200": {
                        "description": "Lista paginada de clientes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Client"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Cria um novo cliente",
                "description": "Store a newly created resource in storage.",
                "operationId": "f909473745c1a1d469e0c5a5b54682b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClientCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/clients/{id}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Mostra detalhes de um cliente",
                "description": "Display the specified resource.",
                "operationId": "feff36e8c89b21277ad216c29ca9ca4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do cliente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Clients"
                ],
                "summary": "Atualiza um cliente existente",
                "description": "Update the specified resource in storage.",
                "operationId": "4cdaf0a748c704e6f11a98f7bb6eb5a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClientUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clients"
                ],
                "summary": "Deleta um cliente",
                "description": "Remove the specified resource from storage.",
                "operationId": "5b6d3bdaf49885db3b7b4b23aed2156b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Cliente deletado com sucesso"
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Client": {
                "required": [
                    "id",
                    "name",
                    "email",
                    "phone",
                    "birthDate",
                    "address",
                    "neighborhood",
                    "zipCode"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "João Silva"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 255,
                        "example": "joao.silva@email.com"
                    },
                    "phone": {
                        "type": "string",
                        "maxLength": 20,
                        "example": "11999999999"
                    },
                    "birthDate": {
                        "type": "string",
                        "format": "date",
                        "example": "1985-12-25"
                    },
                    "address": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Rua Exemplo, 123"
                    },
                    "complement": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Apartamento 101",
                        "nullable": true
                    },
                    "neighborhood": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Centro"
                    },
                    "zipCode": {
                        "type": "string",
                        "maxLength": 9,
                        "example": "12345-678"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-28T20:00:00Z",
                        "nullable": true
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-29T15:30:00Z",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ClientCreateRequest": {
                "required": [
                    "name",
                    "email",
                    "phone",
                    "birthDate",
                    "address",
                    "neighborhood",
                    "zipCode"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "João Silva"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 255,
                        "example": "joao.silva@email.com"
                    },
                    "phone": {
                        "type": "string",
                        "maxLength": 20,
                        "example": "11999999999"
                    },
                    "birthDate": {
                        "type": "string",
                        "format": "date",
                        "example": "1985-12-25"
                    },
                    "address": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Rua Exemplo, 123"
                    },
                    "complement": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Apartamento 101",
                        "nullable": true
                    },
                    "neighborhood": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Centro"
                    },
                    "zipCode": {
                        "type": "string",
                        "maxLength": 9,
                        "example": "12345-678"
                    }
                },
                "type": "object"
            },
            "ClientUpdateRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "João Silva"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 255,
                        "example": "joao.silva@email.com"
                    },
                    "phone": {
                        "type": "string",
                        "maxLength": 20,
                        "example": "11999999999"
                    },
                    "birthDate": {
                        "type": "string",
                        "format": "date",
                        "example": "1985-12-25"
                    },
                    "address": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Rua Exemplo, 123"
                    },
                    "complement": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Apartamento 101",
                        "nullable": true
                    },
                    "neighborhood": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Centro"
                    },
                    "zipCode": {
                        "type": "string",
                        "maxLength": 9,
                        "example": "12345-678"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}